Adicionando validações

+++ src/Main.java
@@ -8,7 +8,7 @@ public class Main {
         Connection conexao = ConexaoDB.conectar();
         Scanner input = new Scanner(System.in);
         TarefaDAO dao = new TarefaDAO(conexao);
-        Menu.menu();
+
 
         int opcao;
         do {
@@ -20,12 +20,23 @@ public class Main {
             switch (opcao) {
                 case 1: {
                     System.out.println("Qual tarefa você quer adicionar? ");
+
                     String tarefa = input.nextLine();
+
+                    if(tarefa.isEmpty()) {
+                        System.out.println("Opção invalida");
+                        break;
+                    }
+
                     System.out.println("Qual o status da tarefa? ");
                     System.out.print("1.Concluida\n2.Em andamento\n3.Não feita");
                     opcao = input.nextInt();
 
                     String status = "";
+
+
+
+
 
                     if (opcao == 1) {
                         status = "Concluida";
@@ -35,7 +46,7 @@ public class Main {
                         status = "Não feita";
                     }else {
                         System.out.println("Opção invalida");
-                        return;
+                        break;
                     }
 
                     Tarefa t = new Tarefa(tarefa, status);
@@ -54,10 +65,26 @@ public class Main {
                     opcao = input.nextInt();
                     input.nextLine();
 
+                    boolean encontrou = false;
+
+                    for (Tarefa t: tarefas) {
+                        if (t.getId() == opcao) {
+                            encontrou = true;
+                            break;
+                        }
+                    }
+
+                    if(!encontrou) {
+                        System.out.println("ID não encontrado");
+                        break;
+                    }
+
                     Tarefa t = new Tarefa();
                     t.setId(opcao);
 
                     dao.remover(t);
+
+                    break;
 
                 } case 3: {
                     ArrayList<Tarefa> tarefas = dao.listar();
@@ -69,6 +96,19 @@ public class Main {
                     System.out.println("Qual o id da tarefa que você deseja alterar?");
 
                     opcao = input.nextInt();
+
+                    boolean encontrou = false;
+
+                    for (Tarefa t: tarefas) {
+                        if (t.getId() == opcao) {
+                            encontrou = true;
+                        }
+                    }
+
+                    if(!encontrou) {
+                        System.out.println("ID não encontrado");
+                        break;
+                    }
                     input.nextLine();
 
                     System.out.println("Qual o novo status da tarefa?");
@@ -85,7 +125,7 @@ public class Main {
                         status = "Não feita";
                     }else {
                         System.out.println("Opção invalida");
-                        return;
+                        break;
                     }
 
                     Tarefa t = new Tarefa();
@@ -94,13 +134,15 @@ public class Main {
 
                     dao.alterar(t);
 
+                    break;
+
                 } case 4: {
                     ArrayList<Tarefa> tarefas = dao.listar();
 
                     for(Tarefa tarefa: tarefas ) {
                         System.out.println(tarefa.getId()+"."+tarefa);
                     }
-
+                    break;
                 }
 
             }
